requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
opencv-python
pillow
numpy
python-multipart
